refactor connection and gossip
